Refactor: move Monitoring classes into their own file
Part of the ongoing breakup of binding.py
